Moving ECS modules around for a more functionality driven structure. Moved input rebinding to it's own crate as it can be used outside amethyst without any issues, and it's optional
