+++ Cargo.toml
@@ -24,6 +24,7 @@ profiler = ["thread_profiler/thread_profiler"]
 amethyst_assets = { path = "amethyst_assets", version = "0.1.0" }
 amethyst_config = { path = "amethyst_config", version = "0.3.0" }
 amethyst_renderer = { path = "amethyst_renderer", version = "0.5" }
+amethyst_input = { path = "amethyst_input", version = "0.1" }
 cgmath = "0.14"
 cpal = "0.4"
 crossbeam = "0.3"
@@ -76,4 +77,4 @@ name = "assets"
 path = "examples/05_assets/main.rs"
 
 [workspace]
-members = ["amethyst_assets", "amethyst_config", "amethyst_renderer"]
+members = ["amethyst_assets", "amethyst_config", "amethyst_renderer", "amethyst_input"]
