+++ amethyst_input/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "amethyst_input"
+version = "0.1.0"
+authors = ["Rhuagh <seamonr@gmail.com>"]
+description = "Input rebinding "
+
+documentation = "https://www.amethyst.rs/doc/master/amethyst_input/"
+homepage = "https://www.amethyst.rs/"
+repository = "https://github.com/amethyst/amethyst"
+
+license = "MIT/Apache-2.0"
+
+[badges]
+appveyor = { repository = "amethyst/amethyst" }
+travis-ci = { repository = "amethyst/amethyst" }
+
+[dependencies]
+winit = "0.7"
+fnv = "1.0"
+serde = "1"
+serde_derive = "1"
+
+[dependencies.smallvec]
+version = "0.4.2"
+features = ["serde"]
