+++ amethyst_input/src/axis.rs
