+++ amethyst_input/src/button.rs
@@ -1,4 +1,4 @@
-use event::{VirtualKeyCode, MouseButton};
+use winit::{VirtualKeyCode, MouseButton};
 
 use super::local_mouse_button::LocalMouseButton;
 use super::local_virtual_key_code::LocalVirtualKeyCode;
