+++ amethyst_input/src/lib.rs
@@ -1,11 +1,10 @@
-//! Contains structures relevant to input handling.
+extern crate serde;
+extern crate winit;
+extern crate fnv;
+extern crate smallvec;
 
-mod axis;
-mod bindings;
-mod button;
-mod input_handler;
-mod local_mouse_button;
-mod local_virtual_key_code;
+#[macro_use]
+extern crate serde_derive;
 
 pub use self::axis::Axis;
 pub use self::bindings::Bindings;
@@ -13,8 +12,14 @@ pub use self::button::{Button, ButtonState, ChangeState};
 pub use self::button::ButtonState::*;
 pub use self::button::ChangeState::*;
 pub use self::input_handler::InputHandler;
-use event::{VirtualKeyCode, MouseButton};
+use winit::{VirtualKeyCode, MouseButton};
 
+mod axis;
+mod bindings;
+mod button;
+mod input_handler;
+mod local_mouse_button;
+mod local_virtual_key_code;
 
 use std::iter::{Chain, Map, Iterator};
 use std::slice::Iter;
@@ -28,7 +33,7 @@ pub type MouseButtons<'a> = Iter<'a, MouseButton>;
 /// An iterator over buttons
 pub struct Buttons<'a> {
     iterator: Chain<Map<Iter<'a, MouseButton>, fn(&MouseButton) -> Button>,
-                    Map<Iter<'a, VirtualKeyCode>, fn(&VirtualKeyCode) -> Button>>,
+        Map<Iter<'a, VirtualKeyCode>, fn(&VirtualKeyCode) -> Button>>,
 }
 
 impl<'a> Iterator for Buttons<'a> {
