+++ amethyst_input/src/local_virtual_key_code.rs
@@ -1,4 +1,4 @@
-use event::VirtualKeyCode;
+use winit::VirtualKeyCode;
 
 #[derive(Serialize, Deserialize)]
 #[serde(remote = "VirtualKeyCode")]
