+++ examples/04_pong/main.rs
@@ -16,10 +16,11 @@ use amethyst::audio::play::*;
 use amethyst::prelude::*;
 use amethyst::assets::{AssetFuture, BoxedErr};
 use amethyst::ecs::{Component, Fetch, FetchMut, Join, System, VecStorage, WriteStorage};
-use amethyst::ecs::components::*;
-use amethyst::ecs::resources::input::{Bindings, InputHandler};
-use amethyst::ecs::resources::Factory;
+use amethyst::ecs::transform::*;
+use amethyst::ecs::rendering::{Factory, MeshComponent, MaterialComponent};
+use amethyst::ecs::input::{Bindings, InputHandler};
 use amethyst::ecs::systems::{DjSystem, TransformSystem};
+use amethyst::ecs::components::*;
 use amethyst::timing::Time;
 use amethyst_renderer::prelude::*;
 use amethyst_renderer::Config as DisplayConfig;
