+++ examples/05_assets/main.rs
@@ -9,12 +9,11 @@ extern crate rayon;
 use amethyst::{Application, Error, State, Trans};
 use amethyst::assets::{AssetFuture, BoxedErr, Context, Format, Loader, NoError};
 use amethyst::assets::formats::textures::{PngFormat, BmpFormat};
-use amethyst::assets::formats::meshes::ObjFormat;
 use amethyst::config::Config;
 use amethyst::ecs::World;
-use amethyst::ecs::resources::input::InputHandler;
-use amethyst::ecs::resources::Factory;
-use amethyst::ecs::components::*;
+use amethyst::ecs::input::InputHandler;
+use amethyst::ecs::rendering::*;
+use amethyst::ecs::transform::{LocalTransform, Transform};
 use amethyst::prelude::*;
 use amethyst::renderer::{Camera, Rgba, Config as DisplayConfig};
 use amethyst::renderer::prelude::*;
@@ -64,10 +63,7 @@ struct AssetsExample;
 
 impl State for AssetsExample {
     fn on_start(&mut self, engine: &mut Engine) {
-        use amethyst::assets::{AssetPtr, Loader};
         use amethyst::assets::formats::meshes::ObjFormat;
-        use amethyst::ecs::components::MeshComponent;
-        use futures::{Async, Future};
 
         let input = InputHandler::new();
         engine.world.add_resource(input);
@@ -176,10 +172,8 @@ fn main() {
 /// Wrapper around the main, so we can return errors easily.
 fn run() -> Result<(), Error> {
     use amethyst::assets::Directory;
-    use amethyst::ecs::components::{Child, Init, LocalTransform, MeshComponent};
-    use amethyst::ecs::systems::TransformSystem;
     use amethyst::ecs::common::Errors;
-    use std::env::set_var;
+    use amethyst::ecs::transform::{Child, Init, LocalTransform, TransformSystem};
 
     // Add our meshes directory to the asset loader.
     let resources_directory = format!(
@@ -198,8 +192,8 @@ fn run() -> Result<(), Error> {
             .clear_target([0.0, 0.0, 0.0, 1.0], 1.0)
             .with_model_pass(pass::DrawShaded::<PosNormTex>::new()),
     );
-    
-    
+
+
 
     let mut game = Application::build(AssetsExample)
         .expect("Failed to build ApplicationBuilder for an unknown reason.")
@@ -232,7 +226,7 @@ fn initialise_camera(camera: &mut Camera) {
 
 /// Adds lights to the scene.
 fn initialise_lights(world: &mut World) {
-    use amethyst::ecs::components::LightComponent;
+    use amethyst::ecs::rendering::LightComponent;
     use amethyst::renderer::light::PointLight;
 
     let light = PointLight {
@@ -280,7 +274,7 @@ fn load_mesh<F>(engine: &mut Engine, name: &str, f: F) -> AssetFuture<MeshCompon
     where F: Format + 'static,
           F::Data: Into<<MeshContext as Context>::Data>,
 {
-    let mut future = {
+    let future = {
         let loader = engine.world.read_resource::<Loader>();
         loader.load_from::<MeshComponent, _, _, _>(name, f, "resources")
     };
