+++ src/ecs/components/mod.rs
@@ -1,9 +1,4 @@
 //! Standard library of useful components.
 
 mod audio;
-mod rendering;
-mod transform;
-
 pub use self::audio::*;
-pub use self::rendering::*;
-pub use self::transform::*;
