+++ src/ecs/input/mod.rs
@@ -1,0 +1,3 @@
+//! `amethyst` input rebinding module
+
+pub use input::{Bindings, InputHandler};
