+++ src/ecs/rendering/components/light.rs
