+++ src/ecs/rendering/components/material.rs
