+++ src/ecs/rendering/components/mod.rs
