+++ src/ecs/rendering/mod.rs
@@ -1,0 +1,9 @@
+//! `amethyst` rendering ecs module
+
+pub mod components;
+pub mod resources;
+pub mod systems;
+
+pub use self::components::*;
+pub use self::resources::*;
+pub use self::systems::*;
