+++ src/ecs/transform/mod.rs
@@ -1,0 +1,7 @@
+//! `amethyst` transform ecs module
+
+pub mod components;
+pub mod systems;
+
+pub use self::components::*;
+pub use self::systems::*;
